requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
et-xmlfile==1.1.0
    # via openpyxl
faker==24.4.0
    # via robotframework-faker
idna==3.6
    # via requests
openpyxl==3.1.2
    # via robotframework-excellib
prompt-toolkit==3.0.43
    # via robotframework-debuglibrary
psycopg2==2.9.9
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via faker
requests==2.31.0
    # via robotframework-requests
robotframework==7.0
    # via
    #   -r requirements.in
    #   robotframework-databaselibrary
    #   robotframework-debuglibrary
    #   robotframework-excellib
    #   robotframework-faker
    #   robotframework-requests
robotframework-csvlib==1.0.2
    # via -r requirements.in
robotframework-databaselibrary==1.4.4
    # via -r requirements.in
robotframework-debuglibrary==2.5.0
    # via -r requirements.in
robotframework-excellib==2.0.1
    # via robotframework-databaselibrary
robotframework-faker==5.0.0
    # via -r requirements.in
robotframework-requests==0.9.6
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
urllib3==2.2.1
    # via requests
wcwidth==0.2.13
    # via prompt-toolkit
wrapt==1.16.0
    # via robotframework-faker
